:truck: Rename centre/router to centre/routes.
Why? To match the content and export name.
